Move ssr middleware into its own file
